requirements.txt
Django>=4.1
python-dotenv
requests
